remove SAT support, devkit-entities should solve ActorTemplate -> Actor
