--- src/actor/SATActor.js
@@ -1,9 +1,0 @@
-import .ActorTemplate;
-import entities.SAT.Entity as SATEntity;
-
-/**
- * For all documentation see {@link ActorTemplate}
- * @class SATActor
- * @extends {SATEntity}
- */
-exports = ActorTemplate(SATEntity);
